feat: update placeSearch for add score

+++ backend/src/main/java/com/ssafy/api/service/PlaceService.java
@@ -2,11 +2,8 @@ package com.ssafy.api.service;
 
 import com.ssafy.api.request.*;
 import com.ssafy.common.handler.CustomException;
-import com.ssafy.domain.document.PlaceMember;
-import com.ssafy.domain.document.Reservation;
-import com.ssafy.domain.document.Review;
+import com.ssafy.domain.document.*;
 import com.ssafy.domain.entity.Member;
-import com.ssafy.domain.document.Place;
 import com.ssafy.domain.repository.*;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -46,6 +43,67 @@ public class PlaceService {
         this.placeMemberRepository = placeMemberRepository;
     }
 
+    public Place getPlace(Long placeId) {
+        return placeRepository.findByPlaceId(placeId).orElse(null);
+    }
+
+    public Reservation getReservation(String reservationId) {
+        return reservationRepository.findById(reservationId).orElse(null);
+    }
+
+    public Review getReview(String reviewId) {
+        return reviewRepository.findById(reviewId).orElse(null);
+    }
+
+    public List<Review> getReviewList(Long placeId) {
+        return reviewRepository.findByPlaceId(placeId);
+    }
+
+    /**
+     * 해당 시설 좋아요 개수
+     */
+    public long getTotalLikeCnt(Long placeId) {
+        return placeMemberRepository.countByLikeItems(placeId);
+    }
+
+    /**
+     * 전체 합산 점수
+     */
+    public double getTotalScore(List<Review> list) {
+        double total = 0.0;
+
+        for(int i = 0; i < list.size(); i++) {
+            total += list.get(i).getCleanness() + list.get(i).getPlace() + list.get(i).getLocation() + list.get(i).getPrice();
+        }
+
+        // 소수점 둘째 자리까지
+        return ( Math.round( total / (list.size() * 4) ) * 100.0 ) / 100.0;
+    }
+
+    /**
+     * 세부 항목의 총 점수와 전체 합산 점수
+     */
+    public double[] getDetailScore(List<Review> list) {
+        double[] score = new double[5];
+
+        for(int i = 0; i < list.size(); i++) {
+            score[0] += list.get(i).getCleanness();
+            score[1] += list.get(i).getPlace();
+            score[2] += list.get(i).getLocation();
+            score[3] += list.get(i).getPrice();
+        }
+
+        score[4] = score[0] + score[1] + score[2] + score[3];
+        score[4] = ( Math.round( score[4] / (list.size() * 4) ) * 100.0 ) / 100.0;
+
+        score[0] = ( Math.round( score[0] / list.size() ) * 100.0 ) / 100.0;
+        score[1] = ( Math.round( score[1] / list.size() ) * 100.0 ) / 100.0;
+        score[2] = ( Math.round( score[2] / list.size() ) * 100.0 ) / 100.0;
+        score[3] = ( Math.round( score[3] / list.size() ) * 100.0 ) / 100.0;
+
+        return score;
+    }
+
     /**
      * 인증 정보에서 memberId값을 가져와 MongoDB에서 유저 정보를 검색
      */
@@ -60,6 +118,10 @@ public class PlaceService {
      */
     public Place getDetailPlace(Long placeId) {
         Member member = memberService.getMemberFromAuthentication();
+        PlaceDetail placeDetail = placeRepositorySupport.detailPlace(placeId);
+
+        // 유저 최근 본 리스트에 저장, viewCnt 증가, detailScore 담기
+
         return null;
     }
 
@@ -79,6 +141,12 @@ public class PlaceService {
         return null;
     }
 
+    /**
+     * 체육시설 전체 목록
+     * 전체를 조회하여 내부 필드값을 수정이 필요하여 사용
+     */
+//    public void allPlaces() {
+//    }
 
     /**
      * 체육시설 장소를 검색
@@ -112,13 +180,14 @@ public class PlaceService {
         }
 
         // 최종 검색
-        return placeRepositorySupport.searchPlace(pageable, searchInfo);
+        return placeRepositorySupport.searchPlaces(pageable, searchInfo);
     }
 
     /**
      * 체육시설 찜 등록/삭제
      */
     public void likePlace(Long placeId) {
+        Place place = getPlace(placeId);
         PlaceMember placeMember = getPlaceMemberFromAuthentication();
         List<Long> likeList = placeMember.getLikeItems();
 
@@ -128,13 +197,16 @@ public class PlaceService {
             likeList.add(placeId);
 
         placeMemberRepository.save(placeMember);
+
+        place.modifyLikeCnt(getTotalLikeCnt(placeId));
+        placeRepository.save(place);
     }
 
     /**
      * 체육시설 리뷰 등록
      */
     public void postReview(ReviewPostReq reviewInfo) {
-        Reservation reservation = reservationRepository.findById(reviewInfo.getReservationId()).orElse(null);
+        Reservation reservation = getReservation(reviewInfo.getReservationId());
 
         // 예약된 정보가 없으면 리뷰 작성 불가
         if(reservation == null) 
@@ -149,6 +221,12 @@ public class PlaceService {
         // 예약 정보의 유저와 현재 로그인한 유저 정보가 다를 때 리뷰 작성 불가
         if(reservation.getMemberId() != memberId)
             throw new CustomException(FAIL_NOT_EQUAL_MEMBER);
+        
+        Place place = getPlace(reservation.getPlaceId());
+
+        // 체육 시설이 조회가 안되면 예약 정보가 잘못 된 것
+        if(place == null)
+            throw new CustomException(FAIL_RESERVATION_PLACE_NOT_FOUND);
 
         // 만약 시설에 대한 리뷰를 작성했는데 다시 리뷰 작성 요청으로 온다면 수정으로 보내기
         Review review = reviewRepository.findById(reservation.getReviewId()).orElse(null);
@@ -188,15 +266,23 @@ public class PlaceService {
 
         reviewRepository.save(newReview);
 
+        // 예약 Collection에 reviewId 담기
         reservation.writtenReview(newReview.getId());
         reservationRepository.save(reservation);
+
+        // place의 리뷰 개수와 점수 수정
+        List<Review> reviewList = getReviewList(review.getPlaceId());
+        if( !(reviewList == null || reviewList.isEmpty()) )
+            place.modifyReviewScore(reviewList.size(), getTotalScore(reviewList));
+
+        placeRepository.save(place);
     }
 
     /**
      * 체육시설 리뷰 수정
      */
     public void modifyReview(ReviewPutReq reviewInfo) {
-        Review review = reviewRepository.findById(reviewInfo.getReviewId()).orElse(null);
+        Review review = getReview(reviewInfo.getReviewId());
 
         // 리뷰가 존재하지 않으면
         if(review == null)
@@ -210,21 +296,32 @@ public class PlaceService {
      * 체육시설 리뷰 삭제
      */
     public void deleteReview(ReviewDeleteReq reviewInfo) {
-        Review review = reviewRepository.findById(reviewInfo.getReviewId()).orElse(null);
-
         // 리뷰가 존재하지 않으면
+        Review review = getReview(reviewInfo.getReviewId());
         if(review == null)
             throw new CustomException(FAIL_REVIEW_NOT_FOUND);
 
-        Reservation reservation = reservationRepository.findById(review.getReservationId()).orElse(null);
+        // 체육 시설이 조회가 안되면 예약 정보가 잘못 된 것
+        Place place = getPlace(review.getPlaceId());
+        if(place == null)
+            throw new CustomException(FAIL_RESERVATION_PLACE_NOT_FOUND);
 
-        // 예약 정보가 없으면 안됨, 혹시 모를 DB 강제 수정 같은 것에 대비 하여 조건문 추가
-        if(reservation != null) {
-            reservation.deleteReview(); // 리뷰 삭제
-            reservationRepository.save(reservation);
-        }
+        // 예약 정보가 존재하지 않으면
+        Reservation reservation = getReservation(review.getReservationId());
+        if(reservation == null)
+            throw new CustomException(FAIL_RESERVATION_NOT_FOUND);
 
         reviewRepository.delete(review);
+
+        // 지우고 조회해야 함 / place의 리뷰 개수와 점수 수정
+        List<Review> reviewList = getReviewList(review.getPlaceId());
+        if( !(reviewList == null || reviewList.isEmpty()) )
+            place.modifyReviewScore(reviewList.size(), getTotalScore(reviewList));
+
+        placeRepository.save(place);
+
+        reservation.deleteReview(); // 리뷰 삭제
+        reservationRepository.save(reservation);
     }
 
     /**
@@ -234,6 +331,11 @@ public class PlaceService {
         // 예약 하려는 날짜가 현재 날짜보다 이전이면 ( 테스트의 편의를 위해 잠시 막아둠 )
 //        if( reservationInfo.getReservationDt().isBefore(LocalDate.now()) )
 //            throw new CustomException(FAIL_RESERVE_BEFORE_NOW_DATE);
+        
+        // 예약 정보에 담긴 시설 정보가 잘못되면
+        Place place = getPlace(reservationInfo.getPlaceId());
+        if(place == null)
+            throw new CustomException(FAIL_RESERVATION_PLACE_NOT_FOUND);
 
         Long memberId = memberService.getMemberIdFromAuthentication();
 
@@ -243,7 +345,7 @@ public class PlaceService {
                 .isWrittenReview(false)
                 .reviewId("")
 //                .createDt(LocalDateTime.now()) // throw Exception 주석 풀면서 같이 풀고 아래 메서드 지울 것!
-                .createDt(reservationInfo.getReservationDt().atTime(0, 00))
+                .createDt(reservationInfo.getReservationDt().atTime(0, 0))
                 .reservationDt(reservationInfo.getReservationDt())
                 .time(reservationInfo.getTime())
                 .price(reservationInfo.getPrice())
@@ -256,7 +358,7 @@ public class PlaceService {
      * 체육시설 예약 취소
      */
     public void cancelPlace(ReservationDeleteReq reservationInfo) {
-        Reservation reservation = reservationRepository.findById(reservationInfo.getReservationId()).orElse(null);
+        Reservation reservation = getReservation(reservationInfo.getReservationId());
 
         // 예약된 정보가 없으면 리뷰 작성 불가
         if(reservation == null)
